backend/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "worksla-backend"
version = "1.0.0"
description = "OpenProject SLA Reporting System Backend"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.0.1",
    "python-multipart==0.0.6",
    "sqlalchemy==2.0.25",
    "psycopg2-binary==2.9.9",
    "alembic==1.13.1",
    "pydantic==2.5.3",
    "pydantic-settings==2.1.0",
    "python-dotenv==1.0.1",
    "requests==2.31.0",
    "urllib3==2.1.0",
    "redis==5.0.1",
    "celery==5.3.6",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.23.3",
    "httpx==0.26.0",
    "black==24.1.1",
    "flake8==7.0.0",
    "mypy==1.8.0",
]
